Unified GLTF viewer and Macha cameras into one
