+++ macha/src/main.rs
@@ -1,3 +1,5 @@
+mod utils;
+
 mod compute_shader_test;
 mod editor;
 mod gltf_loader;
@@ -40,5 +42,5 @@ fn main() {
         .with_dimensions(1280, 720)
         .with_application_name("Macha")
         .with_application_version(0, 1, 0)
-        .build_and_run_inplace::<GLTFViewerState, ()>(());
+        .build_and_run_inplace::<MachaState, ()>(());
 }
